requirements.txt
anyio==3.6.1
autopep8==1.7.0
black==22.8.0
click==8.1.3
colorama==0.4.5
fastapi==0.85.0
greenlet==1.1.3
h11==0.13.0
idna==3.4
mangum==0.15.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
pydantic==1.10.2
PyMySQL==1.0.2
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.3
